requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-mptt==0.13.3
djangorestframework==3.12.4
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
